// arcadia.backoffice.fe/src/modules/reports/components/ReportsGroupByField/types.ts
export enum ActivityReportGroupingKeys {
  operator = 'operator',
  site = 'site',
  group = 'group',
  machine = 'machine',
  month = 'month',
  denomination = 'denomination',
  day = 'day',
}

export enum DisputesReportGroupingKeys {
  operator = 'operator',
  player = 'player',
  status = 'status',
  month = 'month',
  day = 'day',
}

export enum FunnelReportGroupingKeys {
  operator = 'operator',
  site = 'site',
  group = 'group',
  machine = 'machine',
  month = 'month',
  denomination = 'denomination',
  day = 'day',
}

export enum MachineStatusReportGroupingKeys {
  operator = 'operator',
  site = 'site',
  group = 'group',
  machine = 'machine',
  month = 'month',
  denomination = 'denomination',
  day = 'day',
  status = 'status',
}

export enum PlayerBlocksReportGroupingKeys {
  operator = 'operator',
  day = 'day',
  month = 'month',
  reason = 'reason',
}

export enum PlayerStatsReportGroupingKeys {
  player = 'player',
}

export enum RetentionReportGroupingKeys {
  operator = 'operator',
  month = 'month',
  denomination = 'denomination',
  day = 'day',
}

export enum RevenueReportGroupingKeys {
  operator = 'operator',
  site = 'site',
  group = 'group',
  machine = 'machine',
  month = 'month',
  denomination = 'denomination',
  day = 'day',
  currency = 'currency',
}

export enum VouchersReportGroupingKeys {
  operator = 'operator',
  day = 'day',
  month = 'month',
  denomination = 'denomination',
  player = 'player',
}